test(@angular/cli): move `completion` command test in own directory
Just re-organizing a bit
