+++ tests/legacy-cli/e2e/tests/commands/completion/completion.ts
@@ -1,14 +1,13 @@
-import { execFile } from 'child_process';
 import { promises as fs } from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import { getGlobalVariable } from '../../utils/env';
+import { getGlobalVariable } from '../../../utils/env';
 import {
   execAndCaptureError,
   execAndWaitForOutputToMatch,
   execWithEnv,
   silentNpm,
-} from '../../utils/process';
+} from '../../../utils/process';
 
 const testRegistry = getGlobalVariable('package-registry');
